Release 3.12 (#31)
* fix for https://github.com/computational-metabolomics/struct/issues/18 (#19)

- incorrect use of `$assay` replaced with `$data` for DatasetExperiment objects
- specify SummarizedExperiment package as source of `rowData` and `colData` functions

* fix for #20 (#21)

data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* fix for #20 (#23)

data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* Issue 22 seq in (#24)

* add seq_in generics

* add seq_in methods

- added methods for models to get.set seq_in
- added checks for valid input/output names to seq_in and predicted_name methods

* update documentation

* use correct input param id

* update documentation

- use of input_1 instead of value_1 fixed in examples
- corrected rdname for seq_in

* add as.code method (#25)

* add as.code method

produces a string of text that can be used to recreate the model/sequence/iterator

* use correct input parameter names

cmd check fails if examples dont run

* Issue 27 citations (#29)

* add citations and libraries methods for #27

- these methods scan over all inherited objects and generate a complte ist of citations/libraries
- a citations slot has been added to struct_class

* use bibtex for citations

The citations slot is now a list of bibentry objects to provide a more formalised method for storing citations and make them easer to use with bibliography packages etc.

* update imports for utils package

* fix various cmd and bioc check issues

- add importFrom utils citation to namespace
- replace incorrect use of "error" with "stop"
- use first index of class() output in messages

* Update NEWS

* version bump

* fix typo in documentation