updates for bioc_3_13 (#36)
* fix for https://github.com/computational-metabolomics/struct/issues/18 (#19)
- incorrect use of `$assay` replaced with `$data` for DatasetExperiment objects
- specify SummarizedExperiment package as source of `rowData` and `colData` functions

* fix for #20 (#21)
data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* fix for #20 (#23)
data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* Issue 22 seq in (#24)
* add seq_in generics
* add seq_in methods

* update documentation

* add as.code method (#25)
produces a string of text that can be used to recreate the model/sequence/iterator

* use correct input parameter names

* Issue 27 citations (#29)
- these methods scan over all inherited objects and generate a complete list of citations/libraries
- a citations slot has been added to struct_class
* use bibtex for citations
The citations slot is now a list of bibentry objects to provide a more formalised method for storing citations and make them easer to use with bibliography packages etc.

* fix for https://github.com/computational-metabolomics/struct/issues/18 (#19)
- incorrect use of `$assay` replaced with `$data` for DatasetExperiment objects
- specify SummarizedExperiment package as source of `rowData` and `colData` functions

* fix for #20 (#21)
data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* fix for #20 (#23)
data now gets passed through as intended when model_seq is being used. Allows seq_in to be used at the nth step in a sequence

* Update NEWS