Fix as.DatasetExperiment() on SummarizedExperiment(data.frame(..))
Starting with SummarizedExperiment 1.23.2, the SummarizedExperiment()
constructor raises an error if one of the supplied assays has rownames
and/or colnames that don't match those inferred from the supplied 'rowData'
and 'colData' arguments. This change breaks as.DatasetExperiment()
on a SummarizedExperiment object that contains a data.frame assay
(e.g. on SummarizedExperiment(data.frame(x=1:8, y=letters[1:8]))).
This commit repairs as.DatasetExperiment() on such object.